Package was renamed, version number *decreased*
