--- markdown_tool/transformers/__init__.py
@@ -1,8 +1,0 @@
-from markdown_tool.transformers.md.transformer import ArticleTransformer as MarkdownArticleTransformer
-from markdown_tool.transformers.html.transformer import ArticleTransformer as HTMLArticleTransformer
-
-from .helpers import transform_article
-
-
-TRANSFORMERS = [MarkdownArticleTransformer, HTMLArticleTransformer]
-__all__ = [TRANSFORMERS, transform_article]
