+++ markdown_toolset/formatters/__init__.py
@@ -1,9 +1,9 @@
-from markdown_tool.formatters.simple import SimpleFormatter
-from markdown_tool.formatters.html import HTMLFormatter
+from .simple import SimpleFormatter
+from .html import HTMLFormatter
 from .helpers import format_article, get_formatter
 
 try:
-    from markdown_tool.formatters.pdf import PDFFormatter
+    from .pdf import PDFFormatter
 except ModuleNotFoundError:
     PDFFormatter = None
