+++ setup.cfg
@@ -22,7 +22,7 @@ classifiers =
     Topic :: Text Processing :: Markup :: Markdown
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
-    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
